refactor: :art: improve signIn code
