+++ src/webapp/package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@firebase/auth-types": "^0.10.2",
     "@firebase/firestore-types": "^2.1.0",
     "@reduxjs/toolkit": "^1.5.0",
     "@small-ads/ui": "^0.1.0",
