+++ src/webapp/src/auth/types.ts
@@ -1,0 +1,16 @@
+export interface User {
+  uid: string;
+  email: string | null;
+  displayName: string | null;
+}
+export enum AuthStatuses {
+  notLoggedIn,
+  loggedIn,
+  loading,
+}
+
+export interface Login {
+  user: User | null;
+  error?: string | null;
+  authStatus: AuthStatuses;
+}
