+++ src/webapp/src/auth/userSlice.ts
@@ -1,23 +1,6 @@
 import { createSlice } from '@reduxjs/toolkit';
-import { auth } from '../firebase';
-import { signInUser, signOutUser } from './asyncActionsCreators';
-
-export enum AuthStatuses {
-  notLoggedIn,
-  loggedIn,
-  loading,
-}
-
-export interface User {
-  uid: string;
-  email: string | null;
-  displayName: string | null;
-}
-interface Login {
-  user: User | null;
-  error?: string | null;
-  authStatus: AuthStatuses;
-}
+import { AuthStatuses, Login } from './types';
+import { signInUser, signOutUser } from './asyncAuthActions';
 
 const initialState: Login = {
   user: null,
@@ -33,12 +16,11 @@ export const userSlice = createSlice({
       state.authStatus = AuthStatuses.loading;
     });
     builder.addCase(signInUser.rejected, (state, action) => {
-      state.error = action.error?.message;
+      state.error = action.payload?.message || action.error?.message;
       state.authStatus = AuthStatuses.notLoggedIn;
     });
-    builder.addCase(signInUser.fulfilled, (state) => {
-      const { uid, displayName, email } = auth.currentUser!;
-      state.user = { uid, displayName, email };
+    builder.addCase(signInUser.fulfilled, (state, { payload }) => {
+      state.user = payload;
       state.authStatus = AuthStatuses.loggedIn;
     });
     builder.addCase(signOutUser.pending, (state) => {
